Improve exception handling in input switch cases
Refactor switch cases for input detection to handle exceptions more gracefully and not immediately crash the program. 
This would only be happening when no selectable text is in the menu but in that case you should just use console.writeline();.
But who am i to judge when i myself walk as a imperfect man.